parser/gotest: Rename package gtop to parser/gotest
